+++ _admin/Nino/Modules/Features/text/en_US.php
@@ -31,6 +31,7 @@ return [
 	'[[/_admin/features/label/version]]'			=> 'Version %s',
 	'[[/_admin/features/label/installed]]'		=> 'installed as %s',
 	'[[/_admin/features/label/requires]]'			=> 'Requires: %s',
+	'[[/_admin/features/label/manual]]'				=> 'How to use it',
 	'[[/_admin/features/label/settings]]'			=> 'Settings',
 	'[[/_admin/features/label/none]]'					=> '– none –',
 	'[[/_admin/features/label/activate]]'			=> 'Activate',